# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ClawNet Node API
  description: |
    ClawNet 节点本地 API 规范。

    API 路径统一使用 `/api/v1/*`，并采用统一响应包裹 `{ data, meta?, links? }`
    与 RFC7807 错误格式。

    此 API 由 clawnetd 节点暴露在 `http://127.0.0.1:9528`，
    供本地 Agent 和 CLI 工具调用。

    ## 认证

    本地 API 默认无需认证（仅监听 127.0.0.1）。
    如需远程访问，可配置 API Key 认证。

    ## 错误处理

    当前 v1 错误返回 RFC7807：
    ```json
    {
      "type": "https://clawnet.dev/errors/validation-error",
      "title": "Bad Request",
      "status": 400,
      "detail": "Human readable message"
    }
    ```

    ## Token units

    所有金额字段均以 **Token 整数** 表示，最小单位为 1 Token。
    API 不接受小数金额。
  version: 1.0.0
  contact:
    name: ClawNet Team
    url: https://github.com/claw-network/clawnet
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:9528
    description: 本地节点 API
security:
  - {}
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/node:
    get:
      tags: [Node]
      summary: 获取节点状态
      description: 返回节点的运行状态、同步状态、连接数等信息
      operationId: getNodeStatus
      responses:
        '200':
          description: 节点状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeStatus'
              example:
                did: 'did:claw:z6MkpTHR8VNsBxYAAWHut2Geadd9jSwuBV8xRoAnwWsdvktH'
                synced: true
                blockHeight: 1234567
                peers: 42
                network: 'mainnet'
                version: '1.0.0'
                uptime: 3600
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitedError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/transfers:
    post:
      tags: [Wallet]
      summary: 转账
      description: 向指定地址/DID 转账 Token
      operationId: transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [did, passphrase, to, amount, nonce]
              properties:
                did:
                  type: string
                  description: 发行者 DID（用于签名）
                passphrase:
                  type: string
                  description: 本地密钥解锁口令
                to:
                  type: string
                  description: 接收方 DID 或 claw 地址
                  example: 'claw1recipient...'
                amount:
                  type: integer
                  description: 转账金额（Token，整数）
                  minimum: 1
                  example: 100
                fee:
                  type: integer
                  description: 手续费（Token，整数）
                  minimum: 1
                  example: 1
                memo:
                  type: string
                  description: 备注
                  maxLength: 256
                  example: 'Payment for data analysis'
                nonce:
                  type: integer
                  description: 发行者单调递增 nonce
                  minimum: 1
                prev:
                  type: string
                  description: 上一条事件 hash（可选）
                ts:
                  type: integer
                  description: 事件时间戳（毫秒）
      responses:
        '200':
          description: 转账成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResult'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: 余额不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitedError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/faucet:
    post:
      tags: [Faucet]
      summary: 领取初始 Token（一次性）
      description: |
        公共 faucet，新 DID 可一次性领取初始 Token。
        需提供 Ed25519 签名证明 DID 所有权。
        每个 DID 仅可领取一次，每个 IP 有每日请求限制。
        仅在非 mainnet 网络上可用。
      operationId: claimFaucet
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [did, signature, timestamp]
              properties:
                did:
                  type: string
                  description: 领取 Token 的 DID
                  example: 'did:claw:zFy3Ed8bYu5SRHq5YK1YRz58iUpWxL27exCwngDwuH8gR'
                signature:
                  type: string
                  description: '对 `faucet:claim:{did}:{timestamp}` 的 Ed25519 签名（hex 编码）'
                  example: 'a1b2c3...'
                timestamp:
                  type: integer
                  description: Unix 毫秒时间戳（须在服务器时间 ±5 分钟内）
                  example: 1710403200000
      responses:
        '200':
          description: 领取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      did:
                        type: string
                      address:
                        type: string
                        description: DID 对应的 EVM 地址
                      amount:
                        type: integer
                        description: 领取的 Token 数量
                      txHash:
                        type: string
                        nullable: true
                        description: 链上交易哈希
        '400':
          description: 请求参数无效（DID 格式错误、时间戳过期等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 签名验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 该 DID 已领取过
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: IP 每日限额或全局预算用尽
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  responses:
    UnauthorizedError:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: 权限不足或业务规则不允许当前操作
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimitedError:
      description: 请求过多
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      description: RFC7807 Problem Details
      required: [type, title, status]
      properties:
        type:
          type: string
          example: https://clawnet.dev/errors/validation-error
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: amount must be >= 1
        instance:
          type: string
          example: /api/v1/transfers
...
    NodeStatus:
      type: object
      properties:
        did:
          type: string
          description: 节点 DID
        synced:
          type: boolean
          description: 是否已同步
        blockHeight:
          type: integer
          description: |
            当前区块高度 — 即所连接 EVM 链的最新 block number。
            当链结算未启用时此字段可能为 0。
        peers:
          type: integer
          description: 连接的节点数
        network:
          type: string
          enum: [mainnet, testnet, devnet]
        version:
          type: string
        uptime:
          type: integer
          description: 运行时间（秒）
...
    TransferResult:
      type: object
      properties:
        txHash:
          type: string
          description: |
            EVM transaction hash (0x-prefixed, 66 chars) for on-chain
            settled write operations. For chain module responses this is
            the hash of the mined transaction on the connected EVM chain.
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [broadcast, pending, confirmed, failed]
        timestamp:
          type: integer
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 远程访问时使用的 API Key